refactor: 상품 도메인을 분리하라
