+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -1,6 +1,17 @@
 package kitchenpos.domain;
 
-import javax.persistence.*;
+import kitchenpos.product.domain.Product;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.ForeignKey;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import javax.persistence.Transient;
 import java.util.UUID;
 
 @Table(name = "menu_product")
